Modify test command to not build module 'in place'
