+++ README.md
@@ -12,4 +12,4 @@ Make this repository a dependency of your project and automate the process of co
 Because this repository is meant to be used by multiple programming languages, there are no affordances for auto-migrating your database (PRs welcome!). Use [schema.sql](https://github.com/endpoints/fantasy-database/blob/master/schema.sql) as a reference for building migrations, and [data.json](https://github.com/endpoints/fantasy-database/blob/master/data.json) for seeding if you'd like to test in something other than SQLite.
 
 #### Usage notes
-- [Python](https://github.com/endpoints/fantasy-database/blob/master/python/README.md)
+- [Python](python)
